64bit alpha patch :)
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@6760 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/misc.c
@@ -170,7 +170,7 @@ UserCacheDir(void)
 int
 EExit(void *code)
 {
-   int                 exitcode = 0;
+   long                 exitcode = 0;
 
    EDBUG(9, "EExit");
 
@@ -214,7 +214,7 @@ EExit(void *code)
      {
 	int                 i;
 
-	exitcode = (int)code;
+	exitcode = code;
 	SoundExit();
 	if (mustdel)
 	  {
